Cleaned up folder structure
moved custom ui components to their own folder separate from shadcn ui folder and renamed for uniformality
