+++ src/app/products/page.tsx
@@ -2,8 +2,7 @@
 
 import { useEffect, useState, useCallback, Suspense } from "react";
 import { useSearchParams, useRouter } from "next/navigation";
-import Image from "next/image";
-import Link from "next/link";
+import { ProductCard } from '@/components/app-ui/ProductCard'
 import { Product } from "@/types/product"; // Import Product interface from centralized types
 
 // Define an interface for the expected API error response structure
@@ -115,13 +114,13 @@ function ProductsPageContent() {
             }
             router.push(`/products?${newParams.toString()}`);
           }}
-          className="border p-2 rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
+          className="border p-2 rounded focus:outline-none focus:ring-2 focus:ring-blue-500 "
         >
-          <option value=":">Sort by</option>
-          <option value="price:asc">Price (Low to High)</option>
-          <option value="price:desc">Price (High to Low)</option>
-          <option value="inStock:asc">Availability (Out of Stock First)</option>
-          <option value="inStock:desc">Availability (In Stock First)</option>
+          <option value=":" className="dark:text-black">Sort by</option>
+          <option value="price:asc" className="dark:text-black">Price (Low to High)</option>
+          <option value="price:desc" className="dark:text-black">Price (High to Low)</option>
+          <option value="inStock:asc" className="dark:text-black">Availability (Out of Stock First)</option>
+          <option value="inStock:desc" className="dark:text-black">Availability (In Stock First)</option>
         </select>
       </div>
 
@@ -144,41 +143,6 @@ function ProductsPageContent() {
   );
 }
 
-// ProductCard component for better modularity
-function ProductCard({ product }: { product: Product }) {
-  return (
-    <Link href={`/products/${product.id}`} className="block"> {/* Ensure Link is block-level */}
-      <div className="bg-white dark:bg-gray-800 rounded-lg shadow-xl overflow-hidden
-                    transition-all duration-300 hover:shadow-2xl hover:scale-[1.02] cursor-pointer">
-        {product.imageUrl && (
-          <div className="relative w-full h-48 sm:h-56 md:h-64 overflow-hidden"> {/* Responsive height */}
-            <Image
-              src={product.imageUrl}
-              alt={product.name}
-              fill // Use fill for responsive images in a container
-              sizes="(max-width: 640px) 100vw, (max-width: 768px) 50vw, 33vw" // Image optimization
-              className="object-cover transition-transform duration-300 hover:scale-110"
-              priority={false} // Only prioritize LCP images
-            />
-          </div>
-        )}
-        <div className="p-4"> {/* Padding for content inside card */}
-          <h2 className="text-xl font-semibold text-gray-900 dark:text-gray-50 mb-1 truncate">
-            {product.name}
-          </h2>
-          <p className="text-lg font-bold text-blue-600 dark:text-blue-400 mb-2">
-            ${product.price.toFixed(2)}
-          </p>
-          <p className="text-sm">
-            <span className={`font-medium ${product.inStock ? "text-green-600" : "text-red-600"}`}>
-              {product.inStock ? "In Stock" : "Out of Stock"}
-            </span>
-          </p>
-        </div>
-      </div>
-    </Link>
-  );
-}
 
 // The main default export for the page
 export default function ProductsPage() {
