+++ src/components/app-ui/HomePageAnimation.tsx
@@ -1,0 +1,14 @@
+// interface productsI {
+//     I : {
+//         src: string,
+//         name: string
+
+//     }[]
+// }
+// export default function homePageAnimation({ I }: productsI) {
+//     return (
+//         <div className=" ">
+//             test
+//         </div>
+//     )
+// }
