+++ src/components/app-ui/OrderListSection.tsx
@@ -3,7 +3,7 @@
 // It handles loading, error, and empty states.
 
 import { Skeleton } from "@/components/ui/skeleton";
-import { OrderCard } from "@/components/order-card"; // Import the OrderCard component
+import { OrderCard } from "@/components/app-ui/OrderCard"; // Import the OrderCard component
 import { Order } from "@/types/product"; // Import the Order interface
 
 interface OrderListSectionProps {
