+++ src/components/app-ui/ToggleTheme.tsx
@@ -1,0 +1,29 @@
+"use client"
+
+import * as React from "react"
+import { useTheme  } from "next-themes" // Importing to change between dark and light mode
+import { Moon, Sun } from "lucide-react" // Creates moon and sun shape 
+
+export function Toggle() {
+    const { theme, setTheme } = useTheme()
+
+    function toggleTheme() {
+        setTheme(theme === "dark" ? "light" : "dark"); // Switches between dark and light mode 
+    };
+
+    return (
+        <button 
+            aria-label="Toggle theme"
+            onClick={toggleTheme}
+            className="inline-flex h-8 w-8 items-center justify-center rounded-md border border-primary text-primary hover:bg-primary/10 transition-colors"> {/* Replaced hardcoded colors with primary */}
+            {/* Increased icon size from text-sm to text-base and scale-50 to scale-75 */}
+            <Sun className="absolute h-[1.2rem] w-[1.2rem] rotate-0 scale-75 transition-all dark:-rotate-90 dark:scale-0" />
+            <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-75" />
+            
+            <span className="sr-only">Toggle theme</span>
+        </button>
+    );
+}
+
+
+
