--- src/components/ui/DLMode.tsx
@@ -1,12 +1,0 @@
-"use client"
-
-import * as React from "react"
-import { ThemeProvider as NextThemesProvider } from "next-themes"
-
-export function Themes({
-    children,
-    ...props
-    
-}: React.ComponentProps<typeof NextThemesProvider>) {
-    return <NextThemesProvider { ...props}>{children}</NextThemesProvider>
-}
