--- src/components/ui/homePageAnimation.tsx
@@ -1,14 +1,0 @@
-interface productsI {
-    I : {
-        src: string,
-        name: string
-
-    }[]
-}
-export default function homePageAnimation({ I }: productsI) {
-    return (
-        <div className=" ">
-            test
-        </div>
-    )
-}
